move number of recipes per ingredient to callbacks and components
