+++ src/app.py
@@ -2,11 +2,8 @@ from dash import Dash, html, dcc
 import dash_bootstrap_components as dbc
 
 # Import component functions
-from .components import header, footer, ingredient_icons, ingredient_filter, distribution_recipe_ratings, number_of_recipes, average_rating #, number_of_recipes_per_ingredient, recipes_and_complexity
+from .components import header, footer, ingredient_icons, ingredient_filter, distribution_recipe_ratings, number_of_recipes, average_rating, number_of_recipes_per_ingredient #, recipes_and_complexity
 from . import callbacks
-# from .number_of_recipes import number_of_recipes
-# from .average_rating import average_rating
-from .number_of_recipes_per_ingredient import number_of_recipes_per_ingredient
 from .recipes_and_complexity import recipes_and_complexity, update_recipe_list
 from os import path as os_path
